Adding from Anoud Branch

+++ OrderProcessingSystem/project01/Interfaces/IPayable.cs
@@ -8,6 +8,6 @@ namespace OrderProcessingSystem.Interfaces
 {
     public interface IPayable
     {
-        void Pay(double amount);
+        void Pay(decimal amount);
     }
 }
