+++ OrderProcessingSystem/project01/Models/ElectronicsProduct.cs
@@ -8,18 +8,14 @@ namespace OrderProcessingSystem.Models
 {
     public class ElectronicsProduct : Product
     {
-        public ElectronicsProduct(string id,
-                                  string name,
-                                  double price,
-                                  int stockQuantity)
+        public ElectronicsProduct(string id, string name, decimal price, int stockQuantity)
             : base(id, name, price, stockQuantity)
         {
         }
 
-        // Override Method
-        public override double CalculateDiscount()
+        public override decimal CalculateDiscount()
         {
-            return Price * 0.10;
+            return Price * 0.10m;
         }
     }
 }
