+++ OrderProcessingSystem/project01/Models/Payment.cs
@@ -4,7 +4,6 @@ namespace OrderProcessingSystem.Models
 {
     public abstract class Payment : IPayable
     {
-        // Abstract Method
-        public abstract void Pay(double amount);
+        public abstract void Pay(decimal amount);
     }
 }
